generator/requirements.txt
kafka-python
pandas
httpx